+++ src/logging/logging-client.ts
@@ -1,6 +1,10 @@
-import { EntitlementsResult, OpaResponse } from '../types';
+import { EntitlementsResult, OpaResponse, RequestContext, SubjectContext } from '../types';
 
 export interface LoggingClient {
-	log(queryResult: OpaResponse<EntitlementsResult>): void | Promise<void>;
+	log(
+		subjectContext: SubjectContext,
+		requestContext: RequestContext,
+		queryResult: OpaResponse<EntitlementsResult>
+	): void | Promise<void>;
 	error(error: unknown): void | Promise<void>;
 }