+++ src/opa-queries/opa-query.client.ts
@@ -4,6 +4,8 @@ import { FeaturesOpaQuery } from './features-opa-query';
 import { AxiosInstance } from 'axios';
 import { PermissionsOpaQuery } from './permissions-opa-query';
 import { RoutesOpaQuery } from './routes-opa-query';
+import { FGAOpaQuery } from './fga-opa-query';
+import { CompositeFGAOpaQuery } from './composite-fga-query';
 
 export class OpaQueryClient extends EntitlementsOpaQuery {
 	private readonly strategy: Record<RequestContextType, EntitlementsOpaQuery>;
@@ -13,7 +15,9 @@ export class OpaQueryClient extends EntitlementsOpaQuery {
 		this.strategy = {
 			[RequestContextType.Feature]: new FeaturesOpaQuery(pdpHost, axiosInstance),
 			[RequestContextType.Permission]: new PermissionsOpaQuery(pdpHost, axiosInstance),
-			[RequestContextType.Route]: new RoutesOpaQuery(pdpHost, axiosInstance)
+			[RequestContextType.Route]: new RoutesOpaQuery(pdpHost, axiosInstance),
+			[RequestContextType.Entity]: new FGAOpaQuery(pdpHost, axiosInstance),
+			[RequestContextType.Composite]: new CompositeFGAOpaQuery(pdpHost, axiosInstance)
 		};
 	}