+++ src/types/entitlements-query.ts
@@ -1,14 +1,18 @@
 import {
 	FeatureEntitlementsContext,
 	PermissionsEntitlementsContext,
-	RouteEntitlementsContext
+	RouteEntitlementsContext,
+	EntityEntitlementsContext,
+	CompositeEntitlementsContext
 } from './request-context';
 import { SubjectContext } from './subject-context';
 
 export type EntitlementsQueryRequestContext =
 	| Omit<FeatureEntitlementsContext, 'type'>
 	| Omit<PermissionsEntitlementsContext, 'type'>
-	| Omit<RouteEntitlementsContext, 'type'>;
+	| Omit<RouteEntitlementsContext, 'type'>
+	| Omit<EntityEntitlementsContext, 'type'>
+	| Omit<CompositeEntitlementsContext, 'type'>;
 
 export interface EntitlementsQuery {
 	subjectContext: SubjectContext;