+++ src/types/entitlements-result.ts
@@ -1,7 +1,5 @@
-import { EntitlementJustifications } from './entitlements-justifications.enum';
-
 export type EntitlementsResult = {
 	result?: boolean;
-	justification?: EntitlementJustifications;
+	justification?: string;
 	monitoring?: true;
 };