+++ src/types/request-context-type.enum.ts
@@ -1,5 +1,7 @@
 export enum RequestContextType {
 	Feature = 'feature',
 	Permission = 'permission',
-	Route = 'route'
+	Route = 'route',
+	Entity = 'entity',
+	Composite = 'composite'
 }