+++ src/types/request-context.ts
@@ -1,5 +1,7 @@
 import { RequestContextType } from './request-context-type.enum';
 
+type Nullable<T> = { [K in keyof T]: T[K] | null };
+
 export interface FeatureEntitlementsContext {
 	type: RequestContextType.Feature;
 	featureKey: string;
@@ -16,4 +18,22 @@ export interface RouteEntitlementsContext {
 	path: string;
 }
 
-export type RequestContext = FeatureEntitlementsContext | PermissionsEntitlementsContext | RouteEntitlementsContext;
+export interface EntityEntitlementsContext {
+	type: RequestContextType.Entity;
+	entityType: string;
+	key: string;
+	action: string;
+}
+
+export interface CompositeEntitlementsContext {
+	type: RequestContextType.Composite;
+	[RequestContextType.Permission]: Nullable<PermissionsEntitlementsContext>;
+	[RequestContextType.Entity]: Nullable<EntityEntitlementsContext>;
+}
+
+export type RequestContext =
+	| FeatureEntitlementsContext
+	| PermissionsEntitlementsContext
+	| RouteEntitlementsContext
+	| EntityEntitlementsContext
+	| CompositeEntitlementsContext;